Update to libmseed 2.19.5.

+++ obspy/io/mseed/src/libmseed/unpackdata.c
@@ -3,7 +3,7 @@
  * STEIM2, GEOSCOPE (24bit and gain ranged), CDSN, SRO and DWWSSN
  * encoded data.
  *
- * modified: 2017.053
+ * modified: 2017.283
  ************************************************************************/
 
 #include <memory.h>
@@ -628,7 +628,7 @@ msr_decode_geoscope (char *input, int samplecount, float *output,
       mantissa = sample32.i;
 
       /* Take 2's complement for mantissa for overflow */
-      if (mantissa > MAX24)
+      if ((unsigned long)mantissa > MAX24)
         mantissa -= 2 * (MAX24 + 1);
 
       /* Store */
@@ -851,7 +851,7 @@ msr_decode_sro (int16_t *input, int samplecount, int32_t *output,
     gainrange = (sint & SRO_GAINRANGE_MASK) >> SRO_SHIFT;
 
     /* Take 2's complement for mantissa */
-    if (mantissa > MAX12)
+    if ((unsigned long)mantissa > MAX12)
       mantissa -= 2 * (MAX12 + 1);
 
     /* Calculate exponent, SRO exponent = 0..10 */
@@ -902,7 +902,7 @@ msr_decode_dwwssn (int16_t *input, int samplecount, int32_t *output,
     sample = (int32_t)sint;
 
     /* Take 2's complement for sample */
-    if (sample > MAX16)
+    if ((unsigned long)sample > MAX16)
       sample -= 2 * (MAX16 + 1);
 
     /* Save sample in output array */
